+++ Tests/LegacyUnit/Exception/DatabaseTest.php
@@ -1,7 +1,6 @@
 <?php
 
 use OliverKlee\PhpUnit\TestCase;
-use TYPO3\CMS\Core\Database\DatabaseConnection;
 
 /**
  * Test case.
@@ -11,60 +10,10 @@ use TYPO3\CMS\Core\Database\DatabaseConnection;
 class Tx_Oelib_Tests_LegacyUnit_Exception_DatabaseTest extends TestCase
 {
     /**
-     * @var bool the saved content of $GLOBALS['TYPO3_DB']->debugOutput
-     */
-    private $savedDebugOutput;
-
-    /**
-     * @var bool the saved content of
-     *              $GLOBALS['TYPO3_DB']->store_lastBuiltQuery
-     */
-    private $savedStoreLastBuildQuery;
-
-    protected function setUp()
-    {
-        $this->savedDebugOutput = $GLOBALS['TYPO3_DB']->debugOutput;
-        $this->savedStoreLastBuildQuery = $GLOBALS['TYPO3_DB']->store_lastBuiltQuery;
-
-        $GLOBALS['TYPO3_DB']->debugOutput = false;
-        $GLOBALS['TYPO3_DB']->store_lastBuiltQuery = true;
-    }
-
-    protected function tearDown()
-    {
-        $GLOBALS['TYPO3_DB']->debugOutput = $this->savedDebugOutput;
-        $GLOBALS['TYPO3_DB']->store_lastBuiltQuery = $this->savedStoreLastBuildQuery;
-    }
-
-    /**
      * @test
      */
-    public function messageForInvalidQueryContainsErrorMessageFromDatabase()
+    public function isException()
     {
-        /** @var DatabaseConnection $databaseAdapter */
-        $databaseAdapter = $GLOBALS['TYPO3_DB'];
-        $databaseAdapter->exec_SELECTquery('asdf', 'tx_oelib_test', '');
-        $subject = new \Tx_Oelib_Exception_Database();
-
-        self::assertContains(
-            'asdf',
-            $subject->getMessage()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function messageForInvalidQueryWithLastQueryEnabledContainsLastQuery()
-    {
-        /** @var DatabaseConnection $databaseAdapter */
-        $databaseAdapter = $GLOBALS['TYPO3_DB'];
-        $databaseAdapter->exec_SELECTquery('asdf', 'tx_oelib_test', '');
-        $subject = new \Tx_Oelib_Exception_Database();
-
-        self::assertContains(
-            'SELECT',
-            $subject->getMessage()
-        );
+        self::assertInstanceOf(\Exception::class, new \Tx_Oelib_Exception_Database());
     }
 }